Make regexp for next/previous link language independent
Previously the string "next" & "previous" were hardcoded in the regexp,
this failed if you are using orkut in a language other than English.
(one real example of possible values for previous/next are: anterior/próxima)

So now I try and use another unique property of next & previous links,
viz, presence of > & < characters in the text.